Implement a variety of additional functionality
Benchmarks can now be measured in phases. Each phase consists of a window in which a set of scheduler parameters is applied. For each scheduler config, perf measurements are run for the duration of the phase. At the end of the phase perf metrics are saved on a file. Application-level performance for this specific window is also reported. The global reporting functionality remains the same.